Add support for batch-signing requests
As documented in the Roughtime spec, servers can
batch together requests, only signing the root of a computed Merkle
tree, in order to increase efficiency.

Following the example of the reference Roughtime implementation,
the default batch size is set to 64. However, this value can be changed
in the config.

Two pieces of benchmark infrastructure are added - a simple
"benchmark mode" on the server, and a "stress test mode" on the client.
These features can be used to help pick an optimal batch size for the
server.

In "benchmark mode", the server does not log any requests. Instead,
it prints out the current request processing speed every second. This
helps to keep the output manageable when using the client's "stress
test" mode.

In "stress test mode", the client sends the same message to the server
in a loop. To prevent accidental flooding of the users's local network,
or a remote server, only loopback addresses are supported in this mode.